working displayOnly resources field
