+++ src/schema/mutation/editRecords.mutation.ts
@@ -6,15 +6,15 @@ import {
   GraphQLString,
 } from 'graphql';
 import GraphQLJSON from 'graphql-type-json';
-import { Record, Version, Form } from '@models';
+import { Record, Version, Form, Resource } from '@models';
 import extendAbilityForRecords from '@security/extendAbilityForRecords';
 import {
   transformRecord,
   getOwnership,
+  hasInaccessibleFields,
   // checkRecordValidation,
 } from '@utils/form';
 import { RecordType } from '../types';
-import { inaccessibleFields } from './editRecord.mutation';
 import { logger } from '@lib/logger';
 import { graphQLAuthCheck } from '@schema/shared';
 import { Types } from 'mongoose';
@@ -68,9 +68,13 @@ export default {
       });
       for (const record of oldRecords) {
         const ability = await extendAbilityForRecords(user, record.form);
+        const parentResource: Resource = await Resource.findById(
+          record.form.resource,
+          'fields'
+        );
         if (
           ability.can('update', record) &&
-          inaccessibleFields(record, args.data, ability).length === 0
+          !hasInaccessibleFields(record, args.data, ability, parentResource)
         ) {
           // const validationErrors = checkRecordValidation(
           //   record,
